# Holds memory/func defs, and validated elements as values in wasm.Module