--- a/internal/wasm/function_definition_test.go
+++ b/internal/wasm/function_definition_test.go
@@ -13,7 +13,7 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 	tests := []struct {
 		name            string
 		m               *Module
-		expected        []*FunctionDefinition
+		expected        []FunctionDefinition
 		expectedImports []api.FunctionDefinition
 		expectedExports map[string]api.FunctionDefinition
 	}{
@@ -43,7 +43,7 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 					ResultNames:   IndirectNameMap{{Index: Index(0), NameMap: NameMap{{Index: Index(0), Name: "y"}}}},
 				},
 			},
-			expected: []*FunctionDefinition{
+			expected: []FunctionDefinition{
 				{
 					index:       0,
 					name:        "fn",
@@ -79,7 +79,7 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 					{Params: []ValueType{ValueTypeF64, ValueTypeF32}, Results: []ValueType{ValueTypeI64}},
 				},
 			},
-			expected: []*FunctionDefinition{
+			expected: []FunctionDefinition{
 				{
 					index:       0,
 					debugName:   ".$0",
@@ -140,11 +140,12 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 					{Params: []ValueType{ValueTypeF64, ValueTypeF32}, Results: []ValueType{ValueTypeI64}},
 				},
 			},
-			expected: []*FunctionDefinition{
+			expected: []FunctionDefinition{
 				{
 					index:       0,
 					debugName:   ".$0",
-					importDesc:  &[2]string{"", ""},
+					importDesc:  [2]string{"", ""},
+					isImport:    true,
 					exportNames: []string{"imported_function"},
 					funcType:    &FunctionType{Params: []ValueType{ValueTypeF64, ValueTypeF32}, Results: []ValueType{ValueTypeI64}},
 				},
@@ -165,7 +166,8 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 				&FunctionDefinition{
 					index:       0,
 					debugName:   ".$0",
-					importDesc:  &[2]string{"", ""},
+					importDesc:  [2]string{"", ""},
+					isImport:    true,
 					exportNames: []string{"imported_function"},
 					funcType:    &FunctionType{Params: []ValueType{ValueTypeF64, ValueTypeF32}, Results: []ValueType{ValueTypeI64}},
 				},
@@ -174,7 +176,8 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 				"imported_function": &FunctionDefinition{
 					index:       0,
 					debugName:   ".$0",
-					importDesc:  &[2]string{"", ""},
+					importDesc:  [2]string{"", ""},
+					isImport:    true,
 					exportNames: []string{"imported_function"},
 					funcType:    &FunctionType{Params: []ValueType{ValueTypeF64, ValueTypeF32}, Results: []ValueType{ValueTypeI64}},
 				},
@@ -208,8 +211,8 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 				FunctionSection: []Index{0, 0, 0, 0, 0},
 				CodeSection:     []*Code{nopCode, nopCode, nopCode, nopCode, nopCode},
 			},
-			expected: []*FunctionDefinition{
-				{moduleName: "module", index: 0, debugName: "module.$0", importDesc: &[2]string{"i", "f"}, funcType: v_v},
+			expected: []FunctionDefinition{
+				{moduleName: "module", index: 0, debugName: "module.$0", importDesc: [2]string{"i", "f"}, isImport: true, funcType: v_v},
 				{moduleName: "module", index: 1, debugName: "module.$1", funcType: v_v},
 				{moduleName: "module", index: 2, debugName: "module.two", funcType: v_v, name: "two"},
 				{moduleName: "module", index: 3, debugName: "module.$3", funcType: v_v},
@@ -217,7 +220,7 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 				{moduleName: "module", index: 5, debugName: "module.five", funcType: v_v, name: "five"},
 			},
 			expectedImports: []api.FunctionDefinition{
-				&FunctionDefinition{moduleName: "module", index: 0, debugName: "module.$0", importDesc: &[2]string{"i", "f"}, funcType: v_v},
+				&FunctionDefinition{moduleName: "module", index: 0, debugName: "module.$0", importDesc: [2]string{"i", "f"}, isImport: true, funcType: v_v},
 			},
 			expectedExports: map[string]api.FunctionDefinition{},
 		},