--- a/internal/wasm/function_definition_test.go
+++ b/internal/wasm/function_definition_test.go
@@ -8,6 +8,11 @@ import (
 )
 
 func TestModule_BuildFunctionDefinitions(t *testing.T) {
+	imp := &Import{
+		Type:     ExternTypeFunc,
+		DescFunc: 2, // Index of type.
+	}
+
 	nopCode := &Code{Body: []byte{OpcodeEnd}}
 	fn := func(uint32) uint32 { return 1 }
 	tests := []struct {
@@ -123,10 +128,7 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 		{
 			name: "with imports",
 			m: &Module{
-				ImportSection: []Import{{
-					Type:     ExternTypeFunc,
-					DescFunc: 2, // Index of type.
-				}},
+				ImportSection: []Import{*imp},
 				ExportSection: []Export{
 					{Name: "imported_function", Type: ExternTypeFunc, Index: 0},
 					{Name: "function_index=1", Type: ExternTypeFunc, Index: 1},
@@ -144,8 +146,7 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 				{
 					index:       0,
 					debugName:   ".$0",
-					importDesc:  [2]string{"", ""},
-					isImport:    true,
+					importDesc:  imp,
 					exportNames: []string{"imported_function"},
 					funcType:    &FunctionType{Params: []ValueType{ValueTypeF64, ValueTypeF32}, Results: []ValueType{ValueTypeI64}},
 				},
@@ -166,8 +167,7 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 				&FunctionDefinition{
 					index:       0,
 					debugName:   ".$0",
-					importDesc:  [2]string{"", ""},
-					isImport:    true,
+					importDesc:  imp,
 					exportNames: []string{"imported_function"},
 					funcType:    &FunctionType{Params: []ValueType{ValueTypeF64, ValueTypeF32}, Results: []ValueType{ValueTypeI64}},
 				},
@@ -176,8 +176,7 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 				"imported_function": &FunctionDefinition{
 					index:       0,
 					debugName:   ".$0",
-					importDesc:  [2]string{"", ""},
-					isImport:    true,
+					importDesc:  imp,
 					exportNames: []string{"imported_function"},
 					funcType:    &FunctionType{Params: []ValueType{ValueTypeF64, ValueTypeF32}, Results: []ValueType{ValueTypeI64}},
 				},
@@ -212,7 +211,7 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 				CodeSection:     []*Code{nopCode, nopCode, nopCode, nopCode, nopCode},
 			},
 			expected: []FunctionDefinition{
-				{moduleName: "module", index: 0, debugName: "module.$0", importDesc: [2]string{"i", "f"}, isImport: true, funcType: v_v},
+				{moduleName: "module", index: 0, debugName: "module.$0", importDesc: &Import{Module: "i", Name: "f"}, funcType: v_v},
 				{moduleName: "module", index: 1, debugName: "module.$1", funcType: v_v},
 				{moduleName: "module", index: 2, debugName: "module.two", funcType: v_v, name: "two"},
 				{moduleName: "module", index: 3, debugName: "module.$3", funcType: v_v},
@@ -220,7 +219,7 @@ func TestModule_BuildFunctionDefinitions(t *testing.T) {
 				{moduleName: "module", index: 5, debugName: "module.five", funcType: v_v, name: "five"},
 			},
 			expectedImports: []api.FunctionDefinition{
-				&FunctionDefinition{moduleName: "module", index: 0, debugName: "module.$0", importDesc: [2]string{"i", "f"}, isImport: true, funcType: v_v},
+				&FunctionDefinition{moduleName: "module", index: 0, debugName: "module.$0", importDesc: &Import{Module: "i", Name: "f"}, funcType: v_v},
 			},
 			expectedExports: map[string]api.FunctionDefinition{},
 		},