--- a/internal/wasm/memory_definition_test.go
+++ b/internal/wasm/memory_definition_test.go
@@ -11,7 +11,7 @@ func TestModule_BuildMemoryDefinitions(t *testing.T) {
 	tests := []struct {
 		name            string
 		m               *Module
-		expected        []*MemoryDefinition
+		expected        []MemoryDefinition
 		expectedImports []api.MemoryDefinition
 		expectedExports map[string]api.MemoryDefinition
 	}{
@@ -31,7 +31,7 @@ func TestModule_BuildMemoryDefinitions(t *testing.T) {
 		{
 			name:            "defines memory{0,}",
 			m:               &Module{MemorySection: &Memory{Min: 0}},
-			expected:        []*MemoryDefinition{{index: 0, memory: &Memory{Min: 0}}},
+			expected:        []MemoryDefinition{{index: 0, memory: &Memory{Min: 0}}},
 			expectedExports: map[string]api.MemoryDefinition{},
 		},
 		{
@@ -44,7 +44,7 @@ func TestModule_BuildMemoryDefinitions(t *testing.T) {
 				GlobalSection: []Global{{}},
 				MemorySection: &Memory{Min: 2, Max: 3, IsMaxEncoded: true},
 			},
-			expected: []*MemoryDefinition{
+			expected: []MemoryDefinition{
 				{
 					index:       0,
 					exportNames: []string{"memory_index=0"},
@@ -72,7 +72,7 @@ func TestModule_BuildMemoryDefinitions(t *testing.T) {
 				},
 				MemorySection: &Memory{Min: 2, Max: 3, IsMaxEncoded: true},
 			},
-			expected: []*MemoryDefinition{
+			expected: []MemoryDefinition{
 				{
 					index:       0,
 					importDesc:  &[2]string{"", ""},