… internal/wasm/table_test.go