```
$ benchstat old.txt new.txt
name                                    old time/op    new time/op    delta
Initialization/interpreter-32             39.2µs ± 4%    38.9µs ± 5%    ~     (p=0.375 n=17+19)
Initialization/interpreter-multiple-32    38.5µs ± 7%    38.1µs ± 6%    ~     (p=0.247 n=19+20)
Initialization/compiler-32                30.4µs ± 4%    30.3µs ± 2%    ~     (p=0.095 n=17+20)
Initialization/compiler-multiple-32       25.6µs ± 5%    25.4µs ± 4%    ~     (p=0.341 n=20+20)
Compilation/with_extern_cache-32           206µs ± 1%     204µs ± 1%  -1.17%  (p=0.000 n=20+20)
Compilation/without_extern_cache-32       6.12ms ± 1%    6.14ms ± 1%    ~     (p=0.127 n=19+20)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-32              136kB ± 0%     136kB ± 0%    ~     (p=0.547 n=20+20)
Initialization/interpreter-multiple-32     137kB ± 0%     137kB ± 0%    ~     (p=0.215 n=18+16)
Initialization/compiler-32                 137kB ± 0%     137kB ± 0%    ~     (all equal)
Initialization/compiler-multiple-32        142kB ± 0%     142kB ± 0%    ~     (p=0.978 n=18+17)
Compilation/with_extern_cache-32          55.6kB ± 0%    55.2kB ± 0%  -0.73%  (p=0.000 n=20+20)
Compilation/without_extern_cache-32       2.00MB ± 0%    2.00MB ± 0%  -0.02%  (p=0.000 n=20+20)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-32               38.0 ± 0%      38.0 ± 0%    ~     (all equal)
Initialization/interpreter-multiple-32      57.0 ± 0%      57.0 ± 0%    ~     (all equal)
Initialization/compiler-32                  38.0 ± 0%      38.0 ± 0%    ~     (all equal)
Initialization/compiler-multiple-32         47.0 ± 0%      47.0 ± 0%    ~     (all equal)
Compilation/with_extern_cache-32           1.08k ± 0%     1.03k ± 0%  -4.26%  (p=0.000 n=20+20)
Compilation/without_extern_cache-32        33.6k ± 0%     33.5k ± 0%  -0.14%  (p=0.000 n=20+20)


```